# gui/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for the MowgliNext autonomous robot mower GUI",
        "title": "MowgliNext GUI API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4200",
    "basePath": "/api",
    "paths": {
...
        "/mowglinext/call/{command}": {
            "post": {
                "description": "call a service",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mowglinext"
                ],
                "summary": "call a service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "command to call, could be: high_level_control, emergency, mow_enabled, start_in_area",
                        "name": "command",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "CallReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OkResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mowglinext/map": {
            "put": {
                "description": "clear the map and insert all provided areas in a single transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mowglinext"
                ],
                "summary": "Delete the current map and replace all areas",
                "parameters": [
                    {
                        "description": "replace map request body",
                        "name": "CallReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mowgli.ReplaceMapReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OkResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "clear the map",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mowglinext"
                ],
                "summary": "clear the map",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OkResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mowglinext/map/area/add": {
            "post": {
                "description": "add a map area",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mowglinext"
                ],
                "summary": "add a map area",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "CallReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mowgli.AddMowingAreaReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OkResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mowglinext/map/docking": {
            "post": {
                "description": "set the docking point",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mowglinext"
                ],
                "summary": "set the docking point",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "CallReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mowgli.SetDockingPointReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.OkResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/mowglinext/publish/{topic}": {
            "get": {
                "description": "publish to a topic",
                "tags": [
                    "mowglinext"
                ],
                "summary": "publish to a topic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "topic to publish to, could be: joy",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/mowglinext/subscribe/{topic}": {
            "get": {
                "description": "subscribe to a topic",
                "tags": [
                    "mowglinext"
                ],
                "summary": "subscribe to a topic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "logical topic key: diagnostics, status, highLevelStatus, gps, pose, imu, ticks, map, path, plan, mowingPath, power, emergency, dockingSensor, lidar",
                        "name": "topic",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "api.OkResponse": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "string"
                }
            }
        },
...
        "geometry.Point": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "geometry.Point32": {
            "type": "object",
            "properties": {
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "geometry.Polygon": {
            "type": "object",
            "properties": {
                "points": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/geometry.Point32"
                    }
                }
            }
        },
        "geometry.Pose": {
            "type": "object",
            "properties": {
                "orientation": {
                    "$ref": "#/definitions/geometry.Quaternion"
                },
                "position": {
                    "$ref": "#/definitions/geometry.Point"
                }
            }
        },
        "geometry.Quaternion": {
            "type": "object",
            "properties": {
                "w": {
                    "type": "number"
                },
                "x": {
                    "type": "number"
                },
                "y": {
                    "type": "number"
                },
                "z": {
                    "type": "number"
                }
            }
        },
        "mowgli.AddMowingAreaReq": {
            "type": "object",
            "properties": {
                "area": {
                    "$ref": "#/definitions/mowgli.MapArea"
                },
                "is_navigation_area": {
                    "type": "boolean"
                }
            }
        },
        "mowgli.MapArea": {
            "type": "object",
            "properties": {
                "area": {
                    "$ref": "#/definitions/geometry.Polygon"
                },
                "is_navigation_area": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "obstacles": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/geometry.Polygon"
                    }
                }
            }
        },
        "mowgli.ReplaceMapArea": {
            "type": "object",
            "properties": {
                "area": {
                    "$ref": "#/definitions/mowgli.MapArea"
                },
                "is_navigation_area": {
                    "type": "boolean"
                }
            }
        },
        "mowgli.ReplaceMapReq": {
            "type": "object",
            "properties": {
                "areas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mowgli.ReplaceMapArea"
                    }
                }
            }
        },
        "mowgli.SetDockingPointReq": {
            "type": "object",
            "properties": {
                "docking_pose": {
                    "$ref": "#/definitions/geometry.Pose"
                }
            }
        },
...
    }
}