# gui/pkg/api/mowglinext.go
package api

import (
	"context"
	"encoding/base64"
	"encoding/json"
	"errors"
	"log"
	"net/http"
	"time"

	"github.com/cedbossneo/mowglinext/pkg/msgs/geometry"
	"github.com/cedbossneo/mowglinext/pkg/msgs/mowgli"
	"github.com/cedbossneo/mowglinext/pkg/types"
	"github.com/docker/distribution/uuid"
	"github.com/gin-gonic/gin"
	"github.com/gorilla/websocket"
)

var upgrader = websocket.Upgrader{
	ReadBufferSize:  1024,
	WriteBufferSize: 1024,
	CheckOrigin:     func(r *http.Request) bool { return true },
}

func MowgliNextRoutes(r *gin.RouterGroup, provider types.IRosProvider) {
	group := r.Group("/mowglinext")
	ServiceRoute(group, provider)
	AddMapAreaRoute(group, provider)
	SetDockingPointRoute(group, provider)
	ClearMapRoute(group, provider)
	ReplaceMapRoute(group, provider)
	SubscriberRoute(group, provider)
	PublisherRoute(group, provider)
}

// AddMapAreaRoute add a map area
//
// @Summary add a map area
// @Description add a map area
// @Tags mowglinext
// @Accept  json
// @Produce  json
// @Param CallReq body mowgli.AddMowingAreaReq true "request body"
// @Success 200 {object} OkResponse
// @Failure 500 {object} ErrorResponse
// @Router /mowglinext/map/area/add [post]
func AddMapAreaRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.POST("/map/area/add", func(c *gin.Context) {
		ctx, cancel := context.WithTimeout(c.Request.Context(), 10*time.Second)
		defer cancel()

		var CallReq mowgli.AddMowingAreaReq
		err := unmarshalROSMessage[*mowgli.AddMowingAreaReq](c.Request.Body, &CallReq)
		if err != nil {
			return
		}
		if CallReq.Area.Obstacles == nil {
			CallReq.Area.Obstacles = []geometry.Polygon{}
		}
		err = provider.CallService(ctx, "/map_server_node/add_area", &CallReq, &mowgli.AddMowingAreaRes{})
		if err != nil {
			c.JSON(500, ErrorResponse{Error: err.Error()})
		} else {
			c.JSON(200, OkResponse{})
		}
	})
}

// ClearMapRoute delete a map area
//
// @Summary clear the map
// @Description clear the map
// @Tags mowglinext
// @Accept  json
// @Produce  json
// @Success 200 {object} OkResponse
// @Failure 500 {object} ErrorResponse
// @Router /mowglinext/map [delete]
func ClearMapRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.DELETE("/map", func(c *gin.Context) {
		ctx, cancel := context.WithTimeout(c.Request.Context(), 10*time.Second)
		defer cancel()

		err := provider.CallService(ctx, "/map_server_node/clear_map", &mowgli.ClearMapReq{}, &mowgli.ClearMapRes{})
		if err != nil {
			c.JSON(500, ErrorResponse{Error: err.Error()})
		} else {
			c.JSON(200, OkResponse{})
		}
	})
}

// ReplaceMapRoute clear the map and insert areas
//
// @Summary Delete the current map and replace all areas
// @Description clear the map and insert all provided areas in a single transaction
// @Tags mowglinext
// @Accept  json
// @Produce  json
// @Param CallReq body mowgli.ReplaceMapReq true "replace map request body"
// @Success 200 {object} OkResponse
// @Failure 500 {object} ErrorResponse
// @Router /mowglinext/map [put]
func ReplaceMapRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.PUT("/map", func(c *gin.Context) {
		ctx, cancel := context.WithTimeout(c.Request.Context(), 30*time.Second)
		defer cancel()

		err := provider.CallService(ctx, "/map_server_node/clear_map", &mowgli.ClearMapReq{}, &mowgli.ClearMapRes{})
		if err != nil {
			c.JSON(500, ErrorResponse{Error: err.Error()})
			return
		}

		var CallReq mowgli.ReplaceMapReq
		err = unmarshalROSMessage[*mowgli.ReplaceMapReq](c.Request.Body, &CallReq)
		if err != nil {
			c.JSON(500, ErrorResponse{Error: err.Error()})
			return
		}
		for _, element := range CallReq.Areas {
			// Ensure Obstacles is an empty slice, not nil — rosbridge
			// rejects null for repeated fields ("msg is not a list type").
			if element.Area.Obstacles == nil {
				element.Area.Obstacles = []geometry.Polygon{}
			}
			areaReq := mowgli.AddMowingAreaReq{
				Area:             element.Area,
				IsNavigationArea: element.IsNavigationArea,
			}
			err = provider.CallService(ctx, "/map_server_node/add_area", &areaReq, &mowgli.AddMowingAreaRes{})
			if err != nil {
				c.JSON(500, ErrorResponse{Error: err.Error()})
				return
			}
		}

		// Persist areas to disk so they survive container restarts
		_ = provider.CallService(ctx, "/map_server_node/save_areas", &mowgli.ClearMapReq{}, &mowgli.ClearMapRes{})

		c.JSON(200, OkResponse{})
	})
}

// SetDockingPointRoute set the docking point
//
// @Summary set the docking point
// @Description set the docking point
// @Tags mowglinext
// @Accept  json
// @Produce  json
// @Param CallReq body mowgli.SetDockingPointReq true "request body"
// @Success 200 {object} OkResponse
// @Failure 500 {object} ErrorResponse
// @Router /mowglinext/map/docking [post]
func SetDockingPointRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.POST("/map/docking", func(c *gin.Context) {
		ctx, cancel := context.WithTimeout(c.Request.Context(), 10*time.Second)
		defer cancel()

		var CallReq mowgli.SetDockingPointReq
		err := unmarshalROSMessage[*mowgli.SetDockingPointReq](c.Request.Body, &CallReq)
		if err != nil {
			return
		}
		err = provider.CallService(ctx, "/map_server_node/set_docking_point", &CallReq, &mowgli.SetDockingPointRes{})
		if err != nil {
			c.JSON(500, ErrorResponse{Error: err.Error()})
		} else {
			c.JSON(200, OkResponse{})
		}
	})
}

// SubscriberRoute subscribe to a topic
//
// @Summary subscribe to a topic
// @Description subscribe to a topic
// @Tags mowglinext
// @Param topic path string true "logical topic key: diagnostics, status, highLevelStatus, gps, pose, imu, ticks, map, path, plan, mowingPath, power, emergency, dockingSensor, lidar"
// @Router /mowglinext/subscribe/{topic} [get]
func SubscriberRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.GET("/subscribe/:topic", func(c *gin.Context) {
		var err error
		topic := c.Param("topic")
		conn, err := upgrader.Upgrade(c.Writer, c.Request, nil)
		if err != nil {
			return
		}
		defer conn.Close()

		var def func()
		switch topic {
		case "diagnostics":
			def, err = subscribe(provider, c, conn, "diagnostics", -1)
		case "status":
			def, err = subscribe(provider, c, conn, "status", -1)
		case "highLevelStatus":
			def, err = subscribe(provider, c, conn, "highLevelStatus", -1)
		case "gps":
			def, err = subscribe(provider, c, conn, "gps", 100)
		case "pose":
			def, err = subscribe(provider, c, conn, "pose", 100)
		case "imu":
			def, err = subscribe(provider, c, conn, "imu", 100)
		case "ticks":
			def, err = subscribe(provider, c, conn, "ticks", 100)
		case "map":
			def, err = subscribe(provider, c, conn, "map", -1)
		case "path":
			def, err = subscribe(provider, c, conn, "path", -1)
		case "plan":
			def, err = subscribe(provider, c, conn, "plan", -1)
		case "power":
			def, err = subscribe(provider, c, conn, "power", -1)
		case "emergency":
			def, err = subscribe(provider, c, conn, "emergency", -1)
		case "dockingSensor":
			def, err = subscribe(provider, c, conn, "dockingSensor", -1)
		case "lidar":
			def, err = subscribe(provider, c, conn, "lidar", 100)
		case "robotDescription":
			def, err = subscribe(provider, c, conn, "robotDescription", -1)
		case "coverageCells":
			def, err = subscribe(provider, c, conn, "coverageCells", -1)
		case "recordingTrajectory":
			def, err = subscribe(provider, c, conn, "recordingTrajectory", -1)
		}
		if err != nil {
			log.Println(err.Error())
			return
		}
		defer def()

		_, _, err = conn.ReadMessage()
		if err != nil {
			c.Error(err)
			return
		}
	})
}

// PublisherRoute publish to a topic
//
// @Summary publish to a topic
// @Description publish to a topic
// @Tags mowglinext
// @Param topic path string true "topic to publish to, could be: joy"
// @Router /mowglinext/publish/{topic} [get]
func PublisherRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.GET("/publish/:topic", func(c *gin.Context) {
		var err error
		conn, err := upgrader.Upgrade(c.Writer, c.Request, nil)
		if err != nil {
			return
		}
		defer conn.Close()
		for {
			_, msg, err := conn.ReadMessage()
			if err != nil {
				c.Error(err)
				break
			}
			var msgObj geometry.Twist
			err = json.Unmarshal(msg, &msgObj)
			if err != nil {
				c.Error(err)
				break
			}
			err = provider.Publish("/cmd_vel", "geometry_msgs/msg/Twist", &msgObj)
			if err != nil {
				c.Error(err)
				break
			}
		}
	})
}

func subscribe(provider types.IRosProvider, c *gin.Context, conn *websocket.Conn, topic string, interval int) (func(), error) {
	id := uuid.Generate()
	uidString := id.String()
	err := provider.Subscribe(topic, uidString, func(msg []byte) {
		if interval > 0 {
			time.Sleep(time.Duration(interval) * time.Millisecond)
		}
		writer, err := conn.NextWriter(websocket.TextMessage)
		if err != nil {
			c.Error(err)
			return
		}
		_, err = writer.Write([]byte(base64.StdEncoding.EncodeToString(msg)))
		if err != nil {
			c.Error(err)
			return
		}
		err = writer.Close()
		if err != nil {
			c.Error(err)
			return
		}
	},
	)
	if err != nil {
		return nil, err
	}
	return func() {
		provider.UnSubscribe(topic, uidString)
	}, nil
}

// ServiceRoute call a service
//
// @Summary call a service
// @Description call a service
// @Tags mowglinext
// @Accept  json
// @Produce  json
// @Param command path string true "command to call, could be: high_level_control, emergency, mow_enabled, start_in_area"
// @Param CallReq body map[string]interface{} true "request body"
// @Success 200 {object} OkResponse
// @Failure 500 {object} ErrorResponse
// @Router /mowglinext/call/{command} [post]
func ServiceRoute(group *gin.RouterGroup, provider types.IRosProvider) {
	group.POST("/call/:command", func(c *gin.Context) {
		command := c.Param("command")
		var err error
		switch command {
		case "high_level_control":
			var CallReq mowgli.HighLevelControlReq
			err = c.BindJSON(&CallReq)
			if err != nil {
				return
			}
			err = provider.CallService(c.Request.Context(), "/behavior_tree_node/high_level_control", &CallReq, &mowgli.HighLevelControlRes{})
		case "emergency":
			var CallReq mowgli.EmergencyStopReq
			err = c.BindJSON(&CallReq)
			if err != nil {
				return
			}
			err = provider.CallService(c.Request.Context(), "/hardware_bridge/emergency_stop", &CallReq, &mowgli.EmergencyStopRes{})
		case "mow_enabled":
			var CallReq mowgli.MowerControlReq
			err = c.BindJSON(&CallReq)
			if err != nil {
				return
			}
			err = provider.CallService(c.Request.Context(), "/hardware_bridge/mower_control", &CallReq, &mowgli.MowerControlRes{})
		case "start_in_area":
			var CallReq mowgli.StartInAreaReq
			err = c.BindJSON(&CallReq)
			if err != nil {
				return
			}
			err = provider.CallService(c.Request.Context(), "/behavior_tree_node/start_in_area", &CallReq, &mowgli.StartInAreaRes{})
		default:
			err = errors.New("unknown command")
		}
		if err != nil {
			c.JSON(500, ErrorResponse{Error: err.Error()})
		} else {
			c.JSON(200, OkResponse{})
		}
	})
}